load as js module
