+++ Documentation/Blazorise.Docs/Pages/Docs/Examples/AnimateScriptsExample.snippet
@@ -1,1 +1,1 @@
-<script src="_content/Blazorise.Animate/blazorise.animate.js?v=2.1.1.0"></script>
+<!-- No script tag is required. Animate imports its JavaScript module automatically. -->
