+++ Documentation/Blazorise.Docs/Pages/Docs/Extensions/Animate/AnimatePage.razor
@@ -59,7 +59,7 @@
 
 <DocsPageSection>
     <DocsPageSectionHeader Title="Static Files">
-        Add <Code>animate.js</Code> to your <Code>index.html</Code> or <Code>_Host.cshtml</Code> file, depending if you're using a Blazor WebAssembly or Blazor Server side project.
+        No script tag is required. The <Code>Animate</Code> component imports its JavaScript module automatically.
     </DocsPageSectionHeader>
     <DocsPageSectionSource Code="AnimateResourcesExample" />
 </DocsPageSection>
