+++ Documentation/Blazorise.Docs/Pages/Docs/Extensions/Animate/Code/AnimateResourcesExampleCode.html
@@ -1,5 +1,5 @@
 <div class="blazorise-codeblock">
 <div class="html"><pre>
-<span class="htmlTagDelimiter">&lt;</span><span class="htmlElementName">script</span> <span class="htmlAttributeName">src</span><span class="htmlOperator">=</span><span class="quot">&quot;</span><span class="htmlAttributeValue">_content/Blazorise.Animate/blazorise.animate.js?v=2.1.1.0</span><span class="quot">&quot;</span><span class="htmlTagDelimiter">&gt;</span><span class="htmlTagDelimiter">&lt;/</span><span class="htmlElementName">script</span><span class="htmlTagDelimiter">&gt;</span>
+<span class="htmlComment">&lt;!-- No script tag is required. Animate imports its JavaScript module automatically. --&gt;</span>
 </pre></div>
 </div>
